scraper/scrap-spotify/pyproject.toml
[project]
name = "scrap-spotify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "spotifyscraper>=2.1.5",
    "yt-dlp>=2026.3.17",
]
